+++ tests/Unit/Messaging/CloudMessageTest.php
@@ -11,7 +11,6 @@ use Kreait\Firebase\Messaging\CloudMessage;
 use Kreait\Firebase\Messaging\FcmOptions;
 use Kreait\Firebase\Messaging\MessageData;
 use Kreait\Firebase\Messaging\MessageTarget;
-use Kreait\Firebase\Messaging\Notification;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
@@ -28,27 +27,14 @@ final class CloudMessageTest extends TestCase
     }
 
     #[Test]
-    public function withChangedTarget(): void
+    public function anEmptyMessageHasNoTarget(): void
     {
-        $original = CloudMessage::withTarget(MessageTarget::TOKEN, 'bar')
-            ->withData(['foo' => 'bar'])
-            ->withNotification(Notification::create('title', 'body'))
-        ;
+        $message = CloudMessage::new();
+        $payload = Json::decode(Json::encode($message), true);
 
-        $changed = $original->withChangedTarget(MessageTarget::TOKEN, 'baz');
-
-        $encodedOriginal = Json::decode(Json::encode($original), true);
-        $encodedOriginal[MessageTarget::TOKEN] = 'baz';
-
-        $encodedChanged = Json::decode(Json::encode($changed), true);
-
-        $this->assertSame($encodedOriginal, $encodedChanged);
-    }
-
-    #[Test]
-    public function anEmptyMessageHasNotTarget(): void
-    {
-        $this->assertFalse(CloudMessage::new()->hasTarget());
+        $this->assertArrayNotHasKey('token', $payload);
+        $this->assertArrayNotHasKey('condition', $payload);
+        $this->assertArrayNotHasKey('topic', $payload);
     }
 
     #[Test]