+++ CHANGELOG.md
@@ -12,6 +12,14 @@ Please read about the future of the Firebase Admin PHP SDK on the
 * The Messaging component doesn't rely on the `CloudMessage` class for message handling anymore. If you provide a
   message as an array and it has an error, the Firebase API will report it. You can still use the `CloudMessage`
   class as a message builder
+* Deprecated the `CloudMessage::withTarget()` method, use the new `toToken()`, `toTopic()` or `toCondition()` methods instead
+
+### Deprecated
+
+* `Kreait\Firebase\Messaging\CloudMessage::withTarget()` 
+* `Kreait\Firebase\Messaging\CloudMessage::withChangedTarget()` 
+* `Kreait\Firebase\Messaging\CloudMessage::target()`
+* `Kreait\Firebase\Messaging\CloudMessage::hasTarget()`
 
 ## [7.15.0] - 2024-09-11